refactor!: remove deprecated GeneratedVaadinButton
